+++ test/inject-loggedUser-koa2-koa-qs.spec.ts
@@ -23,15 +23,9 @@ let requester: ChaiHttp.Agent;
 
 describe('Injecting logged user data - Koa v2.x with Koa-qs', () => {
 
-    before(async () => {
-        if (process.env.NODE_ENV !== 'test') {
-            throw Error(`Running the test suite with NODE_ENV ${process.env.NODE_ENV} may result in permanent data loss. Please use NODE_ENV=test.`);
-        }
+    before(nock.cleanAll);
 
-        nock.cleanAll();
-    });
-
-    it('GET request with a JWT token in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
+    it('GET request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
         const app: Koa = new Koa2();
 
         const logger: Logger = bunyan.createLogger({
@@ -49,13 +43,9 @@ describe('Injecting logged user data - Koa v2.x with Koa-qs', () => {
             .reply(200, constants.USER);
 
         const registerOptions: BootstrapArguments = {
-            info: { name: 'test MS' },
-            swagger: { swagger: 'test swagger' },
             logger,
-            name: 'test MS',
-            baseURL: 'https://controltower.dev',
-            url: 'https://microservice.dev',
-            token: 'ABCDEF',
+            gatewayURL: 'https://controltower.dev',
+            microserviceToken: 'ABCDEF',
             fastlyEnabled: false
         };
 
@@ -87,7 +77,7 @@ describe('Injecting logged user data - Koa v2.x with Koa-qs', () => {
         response.text.should.equal('ok');
     });
 
-    it('DELETE request with a JWT token in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
+    it('DELETE request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
         const app: Koa = new Koa2();
 
         const logger: Logger = bunyan.createLogger({
@@ -105,13 +95,9 @@ describe('Injecting logged user data - Koa v2.x with Koa-qs', () => {
             .reply(200, constants.USER);
 
         const registerOptions: BootstrapArguments = {
-            info: { name: 'test MS' },
-            swagger: { swagger: 'test swagger' },
             logger,
-            name: 'test MS',
-            baseURL: 'https://controltower.dev',
-            url: 'https://microservice.dev',
-            token: 'ABCDEF',
+            gatewayURL: 'https://controltower.dev',
+            microserviceToken: 'ABCDEF',
             fastlyEnabled: false
         };
 
@@ -143,7 +129,7 @@ describe('Injecting logged user data - Koa v2.x with Koa-qs', () => {
         response.text.should.equal('ok');
     });
 
-    it('POST request with a JWT token in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
+    it('POST request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
         const app: Koa = new Koa2();
 
         const logger: Logger = bunyan.createLogger({
@@ -161,13 +147,9 @@ describe('Injecting logged user data - Koa v2.x with Koa-qs', () => {
             .reply(200, constants.USER);
 
         const registerOptions: BootstrapArguments = {
-            info: { name: 'test MS' },
-            swagger: { swagger: 'test swagger' },
             logger,
-            name: 'test MS',
-            baseURL: 'https://controltower.dev',
-            url: 'https://microservice.dev',
-            token: 'ABCDEF',
+            gatewayURL: 'https://controltower.dev',
+            microserviceToken: 'ABCDEF',
             fastlyEnabled: false
         };
 
@@ -202,7 +184,7 @@ describe('Injecting logged user data - Koa v2.x with Koa-qs', () => {
         response.text.should.equal('ok');
     });
 
-    it('PATCH request with a JWT token in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
+    it('PATCH request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
         const app: Koa = new Koa2();
 
         const logger: Logger = bunyan.createLogger({
@@ -220,13 +202,9 @@ describe('Injecting logged user data - Koa v2.x with Koa-qs', () => {
             .reply(200, constants.USER);
 
         const registerOptions: BootstrapArguments = {
-            info: { name: 'test MS' },
-            swagger: { swagger: 'test swagger' },
             logger,
-            name: 'test MS',
-            baseURL: 'https://controltower.dev',
-            url: 'https://microservice.dev',
-            token: 'ABCDEF',
+            gatewayURL: 'https://controltower.dev',
+            microserviceToken: 'ABCDEF',
             fastlyEnabled: false
         };
 
@@ -261,7 +239,7 @@ describe('Injecting logged user data - Koa v2.x with Koa-qs', () => {
         response.text.should.equal('ok');
     });
 
-    it('PUT request with a JWT token in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
+    it('PUT request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
         const app: Koa = new Koa2();
 
         const logger: Logger = bunyan.createLogger({
@@ -279,13 +257,9 @@ describe('Injecting logged user data - Koa v2.x with Koa-qs', () => {
             .reply(200, constants.USER);
 
         const registerOptions: BootstrapArguments = {
-            info: { name: 'test MS' },
-            swagger: { swagger: 'test swagger' },
             logger,
-            name: 'test MS',
-            baseURL: 'https://controltower.dev',
-            url: 'https://microservice.dev',
-            token: 'ABCDEF',
+            gatewayURL: 'https://controltower.dev',
+            microserviceToken: 'ABCDEF',
             fastlyEnabled: false
         };