+++ test/inject-loggedUser-koa2.spec.ts
@@ -22,15 +22,9 @@ let requester: ChaiHttp.Agent;
 
 describe('Injecting logged user data - Koa v2.x', () => {
 
-    before(async () => {
-        if (process.env.NODE_ENV !== 'test') {
-            throw Error(`Running the test suite with NODE_ENV ${process.env.NODE_ENV} may result in permanent data loss. Please use NODE_ENV=test.`);
-        }
+    before(nock.cleanAll);
 
-        nock.cleanAll();
-    });
-
-    it('GET request with a JWT token in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
+    it('GET request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
         const app: Koa = new Koa2();
 
         const logger: Logger = bunyan.createLogger({
@@ -48,13 +42,9 @@ describe('Injecting logged user data - Koa v2.x', () => {
             .reply(200, constants.USER);
 
         const registerOptions: BootstrapArguments = {
-            info: { name: 'test MS' },
-            swagger: { swagger: 'test swagger' },
             logger,
-            name: 'test MS',
-            baseURL: 'https://controltower.dev',
-            url: 'https://microservice.dev',
-            token: 'ABCDEF',
+            gatewayURL: 'https://controltower.dev',
+            microserviceToken: 'ABCDEF',
             fastlyEnabled: false
         };
 
@@ -85,7 +75,7 @@ describe('Injecting logged user data - Koa v2.x', () => {
         response.text.should.equal('ok');
     });
 
-    it('DELETE request with a JWT token in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
+    it('DELETE request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
         const app: Koa = new Koa2();
 
         const logger: Logger = bunyan.createLogger({
@@ -103,13 +93,9 @@ describe('Injecting logged user data - Koa v2.x', () => {
             .reply(200, constants.USER);
 
         const registerOptions: BootstrapArguments = {
-            info: { name: 'test MS' },
-            swagger: { swagger: 'test swagger' },
             logger,
-            name: 'test MS',
-            baseURL: 'https://controltower.dev',
-            url: 'https://microservice.dev',
-            token: 'ABCDEF',
+            gatewayURL: 'https://controltower.dev',
+            microserviceToken: 'ABCDEF',
             fastlyEnabled: false
         };
 
@@ -140,7 +126,7 @@ describe('Injecting logged user data - Koa v2.x', () => {
         response.text.should.equal('ok');
     });
 
-    it('POST request with a JWT token in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
+    it('POST request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
         const app: Koa = new Koa2();
 
         const logger: Logger = bunyan.createLogger({
@@ -158,13 +144,9 @@ describe('Injecting logged user data - Koa v2.x', () => {
             .reply(200, constants.USER);
 
         const registerOptions: BootstrapArguments = {
-            info: { name: 'test MS' },
-            swagger: { swagger: 'test swagger' },
             logger,
-            name: 'test MS',
-            baseURL: 'https://controltower.dev',
-            url: 'https://microservice.dev',
-            token: 'ABCDEF',
+            gatewayURL: 'https://controltower.dev',
+            microserviceToken: 'ABCDEF',
             fastlyEnabled: false
         };
 
@@ -198,7 +180,7 @@ describe('Injecting logged user data - Koa v2.x', () => {
         response.text.should.equal('ok');
     });
 
-    it('PATCH request with a JWT token in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
+    it('PATCH request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
         const app: Koa = new Koa2();
 
         const logger: Logger = bunyan.createLogger({
@@ -216,13 +198,9 @@ describe('Injecting logged user data - Koa v2.x', () => {
             .reply(200, constants.USER);
 
         const registerOptions: BootstrapArguments = {
-            info: { name: 'test MS' },
-            swagger: { swagger: 'test swagger' },
             logger,
-            name: 'test MS',
-            baseURL: 'https://controltower.dev',
-            url: 'https://microservice.dev',
-            token: 'ABCDEF',
+            gatewayURL: 'https://controltower.dev',
+            microserviceToken: 'ABCDEF',
             fastlyEnabled: false
         };
 
@@ -256,7 +234,7 @@ describe('Injecting logged user data - Koa v2.x', () => {
         response.text.should.equal('ok');
     });
 
-    it('PUT request with a JWT token in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
+    it('PUT request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
         const app: Koa = new Koa2();
 
         const logger: Logger = bunyan.createLogger({
@@ -274,13 +252,9 @@ describe('Injecting logged user data - Koa v2.x', () => {
             .reply(200, constants.USER);
 
         const registerOptions: BootstrapArguments = {
-            info: { name: 'test MS' },
-            swagger: { swagger: 'test swagger' },
             logger,
-            name: 'test MS',
-            baseURL: 'https://controltower.dev',
-            url: 'https://microservice.dev',
-            token: 'ABCDEF',
+            gatewayURL: 'https://controltower.dev',
+            microserviceToken: 'ABCDEF',
             fastlyEnabled: false
         };